+++ automl_alex/__init__.py
@@ -2,4 +2,5 @@ from .models import *
 from .automl_alex import *
 from .data_prepare import *
 from .encoders import *
+from .cross_validation import *
 from .__version__ import __version__