+++ automl_alex/models/base.py
@@ -4,7 +4,6 @@ import sys
 import time
 import optuna
 from tqdm import tqdm
-import pickle
 import joblib
 
 import sklearn
@@ -96,11 +95,95 @@ class ModelBase(object):
         """
         raise NotImplementedError("Pure virtual class.")
 
+
+    def predict_or_predict_proba(self, X):
+        """
+        Сheck and if it is possible get predict_proba
+        """
+        if (self.is_possible_predict_proba()) and \
+                (self.type_of_estimator == 'classifier'):
+            predicts = self.predict_proba(X)
+        else:
+            predicts = self.predict(X)
+        return(predicts)
+
+
+    def _is_possible_feature_importance(self):
+        """
+        Return:
+            bool, whether model can predict proba
+        """
+        return False
+
+
+    def get_feature_importance(self, train_x, importance_type='gain',):
+        """
+        Return:
+            list feature_importance
+        """
+        if not self._is_possible_feature_importance(): 
+            raise Exception("Model cannot get feature_importance")
+        raise NotImplementedError("Pure virtual class.")
+
+
+    def score(self, 
+            X_test, 
+            y_test,
+            metric=None,
+            print_metric=False, 
+            metric_round=4, 
+            ):
+        if self.model is None:
+            raise Exception("No fit models")
+
+        if metric is None:
+            if self.type_of_estimator == 'classifier':
+                metric = sklearn.metrics.roc_auc_score
+            elif self.type_of_estimator == 'regression':
+                metric = sklearn.metrics.mean_squared_error
+
+        # Predict
+        if (metric.__name__ in predict_proba_metrics):
+            y_pred_test = self.predict_or_predict_proba(X_test)
+        else:
+            y_pred_test = self.predict(X_test)
+        score = round(metric(y_test, y_pred_test),metric_round)
+
+        if print_metric:
+            print(f'{metric.__name__}: {score}')
+        return(score)
+
+
+    def fit_score(self, 
+            X_train, 
+            y_train, 
+            X_test, 
+            y_test,
+            metric=None,
+            print_metric=False, 
+            metric_round=4, 
+            ):
+        start = time.time()
+        # Fit
+        self.fit(X_train, y_train,)
+
+        total_time_fit = round((time.time() - start),2)
+        if print_metric:
+            print(f'fit time: {total_time_fit} sec')
+
+        # Score
+        score = self.score(X_test, y_test, 
+            metric=metric,
+            print_metric=print_metric, 
+            metric_round=metric_round,
+            )
+        return(score)
+
+
     def y_format(self, y):
         if isinstance(y, pd.DataFrame):
             y = np.array(y[y.columns[0]].values)
         return y
-
 
 
     def cross_validation(
@@ -183,10 +266,10 @@ class ModelBase(object):
             self.fit(X_train=train_x, y_train=train_y,)
 
             # Predict
-            if (metric.__name__ in predict_proba_metrics) and (self.is_possible_predict_proba()):
-                y_pred = self.predict_proba(val_x)
+            if (metric.__name__ in predict_proba_metrics):
+                y_pred = self.predict_or_predict_proba(val_x)
                 if predict:
-                    y_pred_test = self.predict_proba(X_test)
+                    y_pred_test = self.predict_or_predict_proba(X_test)
             else:
                 y_pred = self.predict(val_x)
                 if predict:
@@ -199,6 +282,7 @@ class ModelBase(object):
                 if i == 0:
                     feature_importance_df = self.get_feature_importance(train_x)
                 feature_importance_df['value'] += self.get_feature_importance(train_x)['value']
+            
             if predict:
                 stacking_y_pred_train[valid_idx] += y_pred
                 stacking_y_pred_test += y_pred_test
@@ -368,7 +452,6 @@ class ModelBase(object):
             trial=trial, 
             opt_lvl=self._opt_lvl, 
             len_data=len_data,
-            metric_name=self.metric.__name__,
             )
         return(self)
 
@@ -686,12 +769,16 @@ class ModelBase(object):
             raise Exception('No history to visualize!')
         return(optuna.visualization.plot_slice(self.study, params=params))
     
-    def save(self, name):
+    def save(self, name, verbose=1):
         joblib.dump(self, name+'.pkl')
-        print('Save Model')
+        if verbose>0:
+            print('Save Model')
 
-    def load(self, name):
-        return(joblib.load(name+'.pkl'))
+    def load(self, name,verbose=1):
+        model = joblib.load(name+'.pkl')
+        if verbose>0:
+            print('Load Model')
+        return(model)